SetupNumberGameFragment has a button that starts to NumberGameCardList activity. The fragment includes a number picker which asks user how many large numbers he wants. When the number pickers value changes button's text need to changes so its implemented.
I've impelemented an custom button, NumberGameCardButton. It has two attributes called as value and isPicked. isPicked defines that the card is selected or not and value attribute is represents the value of number-card. There is an important point that setPicked() method. It has an statement for changing background color to identify isSelected or not.

NumberGameCardListFragment had a listener, OnCardClickedListener. Its name changed as NumbersListListener. From now on, when the fragment onStart, it creates an deal and shuffle them with a simple algorithm(randomly swapping in for loop). Every button which values are small number inside the fragment have an listener. The listener changes background if button is selected. When the numbers of selected numbers is 6, global listener(NumbersListListener) tells to NumberGameCardListActivity to start to game with selected numbers.

NumberGameCardListActivity, starts an activity(NumberGameActivty - game interface) when fragment tells to do it. While starting the activity, we're adding a flag to clear top of the back stack because we dont want to when user clicks back button, it should not go previous display. It should go directly to interface that ask  user how many large numbers he wants.

One of my favorite parts in these lines, I've read something about design patterns and I heard what Memento Design Pattern does. I searched a little on the Internet and I decided to implement for my code, not for now. It will be usefull while user is playing a number game, he might wants to undo. For this scenario, I want to save every process on a stack. The code saves current states of numbersList so I can get back easily with Memento.